move ideal population range to advanced features
