+++ app/src/app/components/sidebar/PopulationPanel/PopulationPanel.tsx
@@ -71,7 +71,9 @@ export const PopulationPanel = () => {
         <Heading as="h3" size="3">
           Total population by district
         </Heading>
-        <PopulationPanelOptions chartOptions={chartOptions} setChartOptions={setChartOptions} />
+        <PopulationPanelOptions chartOptions={chartOptions} setChartOptions={setChartOptions} 
+          idealPopulation={idealPopulation}
+        />
       </Flex>
       <ParentSize
         style={{
@@ -108,68 +110,6 @@ export const PopulationPanel = () => {
               ' will appear when all districts are started'
             )}{' '}
           </Text>
-        </Flex>
-      )}
-      {!!idealPopulation && (
-        <Flex direction="row" align="start" gapX="2" pt="2">
-          <Text>
-            Target deviation from ideal
-            <InfoTip tips="maxDeviation" />
-          </Text>
-          <Flex direction="row" align="center" gapX="2" flexGrow={'1'}>
-            <Flex direction="column" flexGrow={'1'}>
-              <TextField.Root
-                placeholder="% Deviation"
-                type="number"
-                max={100}
-                step={0.1}
-                value={chartOptions.popTargetPopDeviationPct || undefined}
-                onChange={e => {
-                  if (e.target.value === '') {
-                    setChartOptions({
-                      popTargetPopDeviation: undefined,
-                      popTargetPopDeviationPct: undefined,
-                    });
-                  } else {
-                    const value = Math.max(0, +e.target.value);
-                    setChartOptions({
-                      popTargetPopDeviation: Math.round((value / 100) * idealPopulation),
-                      popTargetPopDeviationPct: value,
-                    });
-                  }
-                }}
-              >
-                <TextField.Slot side="right">
-                  <IconButton size="1" variant="ghost">
-                    %
-                  </IconButton>
-                </TextField.Slot>
-              </TextField.Root>
-              <Text size="1">Percent</Text>
-            </Flex>
-            <Flex direction="column" flexGrow={'1'}>
-              <TextField.Root
-                placeholder="Pop Deviation"
-                type="number"
-                value={chartOptions.popTargetPopDeviation || undefined}
-                onChange={e => {
-                  if (e.target.value === '') {
-                    setChartOptions({
-                      popTargetPopDeviation: undefined,
-                      popTargetPopDeviationPct: undefined,
-                    });
-                  } else {
-                    const value = Math.max(0, +e.target.value);
-                    setChartOptions({
-                      popTargetPopDeviation: value,
-                      popTargetPopDeviationPct: Math.round((value / idealPopulation) * 10000) / 100,
-                    });
-                  }
-                }}
-              ></TextField.Root>
-              <Text size="1">Population</Text>
-            </Flex>
-          </Flex>
         </Flex>
       )}
     </Flex>
