+++ app/src/app/components/sidebar/PopulationPanel/PopulationPanelOptions.tsx
@@ -1,4 +1,4 @@
-import {CheckboxGroup, Flex, IconButton, Radio, Text} from '@radix-ui/themes';
+import {CheckboxGroup, Flex, IconButton, Radio, Text, TextField} from '@radix-ui/themes';
 import React from 'react'; // Import ParentSize
 import {Popover} from '@radix-ui/themes';
 import {GearIcon} from '@radix-ui/react-icons';
@@ -8,7 +8,8 @@ import {ChartStore} from '@store/chartStore';
 export const PopulationPanelOptions: React.FC<{
   chartOptions: ChartStore['chartOptions'];
   setChartOptions: ChartStore['setChartOptions'];
-}> = ({chartOptions, setChartOptions}) => {
+  idealPopulation?: number;
+}> = ({chartOptions, setChartOptions, idealPopulation}) => {
   return (
     <Popover.Root>
       <Popover.Trigger>
@@ -94,6 +95,69 @@ export const PopulationPanelOptions: React.FC<{
             />
             <Text size={'2'}>Scale bars to current zone population range</Text>
           </Flex>
+
+      {!!idealPopulation && (
+        <Flex direction="column" align="start" gapX="2" pt="2">
+          <Text className="py-2">
+            Target deviation from ideal
+            <InfoTip tips="maxDeviation" />
+          </Text>
+          <Flex direction="row" align="center" gapX="2" flexGrow={'1'}>
+            <Flex direction="column" flexGrow={'1'}>
+              <TextField.Root
+                placeholder="% Deviation"
+                type="number"
+                max={100}
+                step={0.1}
+                value={chartOptions.popTargetPopDeviationPct || undefined}
+                onChange={e => {
+                  if (e.target.value === '') {
+                    setChartOptions({
+                      popTargetPopDeviation: undefined,
+                      popTargetPopDeviationPct: undefined,
+                    });
+                  } else {
+                    const value = Math.max(0, +e.target.value);
+                    setChartOptions({
+                      popTargetPopDeviation: Math.round((value / 100) * idealPopulation),
+                      popTargetPopDeviationPct: value,
+                    });
+                  }
+                }}
+              >
+                <TextField.Slot side="right">
+                  <IconButton size="1" variant="ghost">
+                    %
+                  </IconButton>
+                </TextField.Slot>
+              </TextField.Root>
+              <Text size="1">Percent</Text>
+            </Flex>
+            <Flex direction="column" flexGrow={'1'}>
+              <TextField.Root
+                placeholder="Pop Deviation"
+                type="number"
+                value={chartOptions.popTargetPopDeviation || undefined}
+                onChange={e => {
+                  if (e.target.value === '') {
+                    setChartOptions({
+                      popTargetPopDeviation: undefined,
+                      popTargetPopDeviationPct: undefined,
+                    });
+                  } else {
+                    const value = Math.max(0, +e.target.value);
+                    setChartOptions({
+                      popTargetPopDeviation: value,
+                      popTargetPopDeviationPct: Math.round((value / idealPopulation) * 10000) / 100,
+                    });
+                  }
+                }}
+              ></TextField.Root>
+              <Text size="1">Population</Text>
+            </Flex>
+          </Flex>
+        </Flex>
+      )}
         </Flex>
       </Popover.Content>
     </Popover.Root>
